pyproject.toml
[project]
name = "dex"
version = "0.1.0"
description = "AI Context Manager for AI-augmented development tools"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Dex Contributors" }
]
keywords = ["ai", "context-manager", "claude", "cursor", "codex", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "tomli-w>=1.0.0",  # TOML writing for Codex config
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.4.0",
    "flake8>=7.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

[project.scripts]
dex = "dex.cli.main:app"

[project.urls]
Homepage = "https://github.com/launchcg/dex"
Repository = "https://github.com/launchcg/dex"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dex"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = '''
/(
    \.venv
    | \.git
    | __pycache__
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", ".git", "__pycache__", "build", "dist"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
exclude = [".venv", "build", "dist", "docs/examples"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
